feat(工具集): 天气agent修改为工具
天气agent修改为工具
